refactored util functions for tickets, screenshots into respective packages in the utils directory. Moved DatabaseManager to the db package in utils #11
